feat(ml): add Node2Vec embeddings and bug prediction model
Add ML-based bug prediction using Node2Vec graph embeddings combined
with code metrics to identify high-risk functions.

Key components:
- Node2VecEmbedder: Generate embeddings via Neo4j GDS with configurable
  walk length, return factor (p), and in-out factor (q) parameters
- FalkorDBNode2VecEmbedder: Fallback using Cypher random walks + gensim
- BugPredictor: RandomForest classifier with 138-dimensional features
  (128 Node2Vec + 10 code metrics like complexity, LOC, coupling)
- FeatureExtractor: Combines embeddings with graph-derived metrics
- MLBugDetector: Integration with analysis engine for findings

CLI commands:
- repotoire ml generate-embeddings: Create Node2Vec embeddings
- repotoire ml train-bug-predictor: Train model on labeled data
- repotoire ml predict-bugs: Identify high-risk functions

Also fixes Pydantic 2 Union type validation error in TrainingDataset.

+++ repotoire/ml/training_data.py
@@ -74,7 +74,7 @@ class TrainingDataset(BaseModel):
     repository: str = Field(description="Path to the repository")
     extracted_at: str = Field(description="ISO format extraction timestamp")
     date_range: Tuple[str, str] = Field(description="(start_date, end_date) for commit range")
-    statistics: Dict[str, Union[float, int, str]] = Field(default_factory=dict, description="Dataset statistics")
+    statistics: Dict[str, Union[int, float, str]] = Field(default_factory=dict, description="Dataset statistics")
     version: str = Field(default="1.0.0", description="Dataset format version")
